+++ phy/plot/_vispy_utils.py
@@ -207,7 +207,7 @@ class BaseSpikeVisual(_BakeVisual):
     def spike_ids(self):
         """Spike ids to display."""
         if self._spike_ids is None:
-            self._spike_ids = np.arange(self.n_spikes).astype(np.uint64)
+            self._spike_ids = np.arange(self.n_spikes)
         return self._spike_ids
 
     @spike_ids.setter
